requirements.txt
flask
flask-cors
pandas
python-dotenv
openai
google-genai